+++ frontend/tsconfig.node.tsbuildinfo
@@ -1,0 +1,1 @@
+{"version":"4.9.5"}
