+++ lab-code/backend/config/index.js
@@ -1,0 +1,5 @@
+export const MONGO_URI = process.env.MONGO_URI || 'mongodb://localhost:27017';
+
+export const ENVIRONMENT = process.env.NODE_ENV || 'development';
+
+export const PORT = process.env.PORT || 3000;
