+++ lab-code/backend/models/BaseEntity.js
@@ -1,0 +1,7 @@
+export class BaseEntity {
+  constructor() {
+    this.id = uuidv4();
+    this.createdAt = new Date();
+    this.updatedAt = new Date();
+  }
+}
