+++ lab-code/backend/util/index.js
@@ -1,0 +1,7 @@
+
+export let validate = (county, state) => {
+    if (county === undefined || state === undefined) {
+        return false;
+    }
+    return true;
+}
